requirements.txt
asgiref==3.7.2
boto3==1.24.17
botocore==1.27.17
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.2.3
django-appconf==1.0.5
django-ckeditor==6.4.2
django-cleanup==6.0.0
django-cors-headers==4.6.0
django-imagekit==4.1.0
django-js-asset==2.0.0
django-modeltranslation==0.18.11
django-multiselectfield==0.1.13
django-storages==1.12.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
minify_html==0.10.3
packaging==24.2
pilkit==2.0
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2024.2
PyYAML==6.0.2
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.7.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
